updates to validation, fix error case with legacy fov names

+++ validateBisqueDb.py
@@ -5,51 +5,47 @@
 
 import argparse
 import cellJob
-import dataHandoffSpreadsheetUtils as utils
+import dataHandoffUtils as utils
 import json
 import os
 import re
 import sys
 import uploader.db_api as db_api
 
-def verify_files(thumbs_dir, data_dir, data_subdir, cell_line, f, batchname, jobname):
+def verify_files(thumbs_dir, data_dir, cell_line, f, jobname):
     # check for thumbnail
-    fullf = os.path.join(thumbs_dir, data_subdir, cell_line, f + '.png')
+    fullf = os.path.join(thumbs_dir, cell_line, f + '.png')
     if not os.path.isfile(fullf):
-        print(batchname + ": " + jobname + ": Could not find file: " + fullf)
+        print(jobname + ": Could not find file: " + fullf)
 
     # check for image
-    fullf = os.path.join(data_dir, data_subdir, cell_line, f + '.ome.tif')
+    fullf = os.path.join(data_dir, cell_line, f + '.ome.tif')
     if not os.path.isfile(fullf):
-        print(batchname + ": " + jobname + ": Could not find file: " + fullf)
+        print(jobname + ": Could not find file: " + fullf)
 
     # check for atlas meta
-    fullaj = os.path.join(thumbs_dir, data_subdir, cell_line, f + '_atlas.json')
+    fullaj = os.path.join(thumbs_dir, cell_line, f + '_atlas.json')
     if not os.path.isfile(fullaj):
-        print(batchname + ": " + jobname + ": Could not find file: " + fullaj)
+        print(jobname + ": Could not find file: " + fullaj)
 
     # expect 3 atlas png files
     for i in ['0', '1', '2']:
-        fullat = os.path.join(thumbs_dir, data_subdir, cell_line, f + '_atlas_'+i+'.png')
+        fullat = os.path.join(thumbs_dir, cell_line, f + '_atlas_'+i+'.png')
         if not os.path.isfile(fullat):
-            print(batchname + ": " + jobname + ": Could not find file: " + fullat)
+            print(jobname + ": Could not find file: " + fullat)
 
     # check for image meta
-    fullmj = os.path.join(thumbs_dir, data_subdir, cell_line, f + '_meta.json')
+    fullmj = os.path.join(thumbs_dir, cell_line, f + '_meta.json')
     if not os.path.isfile(fullmj):
-        print(batchname + ": " + jobname + ": Could not find file: " + fullmj)
+        print(jobname + ": Could not find file: " + fullmj)
 
 
-def do_image(row, prefs):
-    batchname = row['source_data']
-    jobname = row['inputFilename']
+def do_image(row, prefs, is_dryrun):
     info = cellJob.CellJob(row)
+    jobname = info.SourceFilename
 
-    imageName = info.cbrCellName
-    segs = row["outputCellSegIndex"]
-    segs = segs.split(";")
-    # get rid of empty strings in segs
-    segs = list(filter(None, segs))
+    imageName = info.FOV_3dcv_Name
+    segs = info.CellId
 
     names = [imageName]
     for seg in segs:
@@ -58,11 +54,7 @@ def do_image(row, prefs):
 
     data_dir = prefs['out_ometifroot']
     thumbs_dir = prefs['out_thumbnailroot']
-    # assume that the file location has same name as this subdir name of where the spreadsheet lives:
-    path_as_list = re.split(r'\\|/', batchname)
-    data_subdir = path_as_list[-3]
-    # data_subdir = '2017_03_08_Struct_First_Pass_Seg'
-    cell_line = 'AICS-' + str(row["cell_line_ID"])
+    cell_line = info.CellLineName
 
     # get associated images from db
     query_result = []
@@ -70,12 +62,11 @@ def do_image(row, prefs):
     for i in xml:
         query_result.append(i.get("name"))
 
-
     names_in_db = []
     # make sure every segmented cell image is in the db
     ids_to_delete = []
     for f in names:
-        expected_relpath = data_subdir + '/' + cell_line + '/' + f + '.ome.tif'
+        expected_relpath = cell_line + '/' + f + '.ome.tif'
 
         found_in_db = False
         for i in xml:
@@ -87,46 +78,47 @@ def do_image(row, prefs):
                     # repeated image in db.
                     imid = i.get("resource_uniq")
                     print("ERROR: DELETING {} : redundant db entry : {} : {} : {}".format(f, imid, imname, imgnameindb))
-                    ids_to_delete.append(imid);
+                    ids_to_delete.append(imid)
                 elif imgnameindb != expected_relpath:
                     print("ERROR path mismatch for " + f + ": db has " + imgnameindb + ' but expected ' + expected_relpath)
                 else:
                     names_in_db.append(imname)
         if not found_in_db:
-            print('ERROR: {} not found in db ( {},{} )'.format(f, batchname, jobname))
+            print('ERROR: {} not found in db ( {} )'.format(f, jobname))
 
 
     for i in xml:
         imname = i.get("name")
+        # assumes 2 digit aics cell line number 'AICS_##-'
         if not imname[:-8] in names:
-            print('ERROR: DELETING {} : not part of data set for {} ( {},{} ) and will be deleted'.format(imname, imageName, batchname, jobname))
+            print('ERROR: DELETING {} : not part of data set for {} ( {} ) and will be deleted'.format(imname, imageName, jobname))
             imid = i.get("resource_uniq")
             ids_to_delete.append(imid);
         else:
             # check pathing.
             imgnameindb = i.get("value")
-            expected_relpath = data_subdir + '/' + cell_line + '/' + imname
+            expected_relpath = cell_line + '/' + imname
             if imgnameindb != expected_relpath:
                 print("ERROR path mismatch for " + f + ": db has " + imgnameindb + ' but expected ' + expected_relpath)
 
     # UNCOMMENT TO DO ACTUAL DELETIONS. DANGER THIS MAY HAVE SIDE EFFECT OF REMOVING OME TIF FILES.
-    for i in ids_to_delete:
-        db_api.DbApi.deleteImage(i)
+    if not is_dryrun:
+        for i in ids_to_delete:
+            db_api.DbApi.deleteImage(i)
 
     for f in names:
-        verify_files(thumbs_dir, data_dir, data_subdir, cell_line, f, batchname, jobname)
+        verify_files(thumbs_dir, data_dir, cell_line, f, jobname)
+
 
 def parse_args():
     parser = argparse.ArgumentParser(description='Process data set defined in csv files, '
                                                  'and set up a job script for each row.'
                                                  'Example: python createJobsFromCSV.py -c -n --dataset 2017_03_08_Struct_First_Pass_Seg')
 
-    # python validateDataHandoff --sheets D:\src\aics\dataset_cellnuc_seg_curated\2017_05_15_tubulin\spreasheets_contourXY
-
     parser.add_argument('prefs', nargs='?', default='prefs.json', help='input prefs')
 
     # sheets replaces input...
-    parser.add_argument('--sheets', help='directory containing *.xlsx', default='')
+    parser.add_argument('--dryrun', help='perform deletions if switched off', default='True')
 
     args = parser.parse_args()
 
@@ -171,16 +163,15 @@ def report_db_stats():
 
 def do_main(args, prefs):
     # Read every .csv file and concat them together
-    data = utils.collect_data_rows(prefs['data_files'], db_path=prefs['imageIDs'])
+    data = utils.collect_data_rows()
+    data = data.to_dict(orient='records')
 
     total_jobs = len(data)
     print('VALIDATING ' + str(total_jobs) + ' JOBS')
     total_cells = 0
     for index, row in enumerate(data):
-        segs = row["outputCellSegIndex"]
-        segs = segs.split(";")
+        segs = row["CellId"]
         # get rid of empty strings in segs
-        segs = list(filter(None, segs))
         total_cells += len(segs)
     print('EXPECTING ' + str(total_cells) + ' SINGLE CELLS')
 
@@ -198,7 +189,7 @@ def do_main(args, prefs):
     for index, row in enumerate(data):
         if index % 100 == 0:
             print(str(index))
-        do_image(row, prefs)
+        do_image(row, prefs, args.dryrun)
 
     print('**DB REPORT**')
     report_db_stats()
